Credit card form WIP
There is some added complexity to this form due to the fact that the vue-credit-card plugin does not offer any way to see if the credit card form is valid, or an easy way to see if the individual fields are valid.  Accessing validity information for individual fields through the $refs object is somewhat effective, but this information is not updated immediately, and so calculating the validity of the form when the input event is fired is not guaranteed to yield correct results.